requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
networkx>=2.6.0
scipy>=1.7.0
pytest>=6.2.5
langchain>=0.0.267
openai>=1.0.0
transformers>=4.30.0
tqdm>=4.62.0
pyyaml>=6.0 